vfs: move the link family into vfs/link.c
Relocate do_link/do_link_at/do_link_atfd out of syscall.c into
vfs/link.c as the vfs_* names, declared in vfs/vfs.h.  The
HAVE_LINK/HAVE_LINKAT guards travel verbatim.  No behavior change.
