+++ vfs/vfs.h
@@ -117,4 +117,9 @@ ssize_t vfs_readlink_atfd(int dfd, const char *name, char *buf, size_t bufsiz);
 ssize_t vfs_readlink(const char *path, char *buf, size_t bufsiz);
 #endif
 
+/* hard links (vfs/link.c). */
+int vfs_link(const char *old_path, const char *new_path);
+int vfs_link_at(const char *old_path, const char *new_path);
+int vfs_link_atfd(int old_dfd, const char *old_name, int new_dfd, const char *new_name, int flags);
+
 #endif /* RSYNC_VFS_H */
